hw1/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.1.19=hecd8cb5_1
cairo=1.16.0=h8023c5d_1
certifi=2020.12.5=py37hecd8cb5_0
cycler=0.10.0=py37_0
ffmpeg=4.0=h01ea3c9_0
fontconfig=2.13.1=ha9ee91d_0
freetype=2.10.4=ha233b18_0
gettext=0.19.8.1=hb0f4f8b_2
glib=2.67.4=hdf23fa2_1
graphite2=1.3.14=h38d11af_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h0a44026_3
intel-openmp=2019.4=233
jasper=2.0.14=h636a363_1
jpeg=9b=he5867d9_2
kiwisolver=1.3.1=py37h23ab428_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h9da4c3f_0
libvpx=1.7.0=h378b8a2_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.10=h7cdb67c_3
lz4-c=1.9.3=h23ab428_0
matplotlib=3.3.4=py37hecd8cb5_0
matplotlib-base=3.3.4=py37h8b3ea08_0
mkl=2019.4=233
mkl-service=2.3.0=py37h9ed2024_0
mkl_fft=1.3.0=py37ha059aab_0
mkl_random=1.1.1=py37h959d312_0
ncurses=6.2=h0a44026_1
numpy=1.19.2=py37h456fd55_0
numpy-base=1.19.2=py37hcfb5961_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1j=h9ed2024_0
pcre=8.44=hb1e8313_0
pillow=8.1.2=py37h5270095_0
pip=21.0.1=py37hecd8cb5_0
pixman=0.40.0=haf1e3a3_0
py-opencv=3.4.2=py37h7c891bd_1
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.7.10=h88f2d9e_0
python-dateutil=2.8.1=pyhd3eb1b0_0
readline=8.1=h9ed2024_0
scipy=1.6.1=py37h2515648_0
setuptools=52.0.0=py37hecd8cb5_0
six=1.15.0=py37hecd8cb5_0
sqlite=3.35.1=hce871da_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py37h9ed2024_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
